Fix spacing issues
There were places where we were missing spaces because IntelliJ's Markdown
renderer did some weird things.

I also added a little text and a few links to expand on a couple of ideas. This include
a mention of the ability to use Google Fonts and a link to info on the fairly recent shift
to more views on phones than on monitors.