Update LABTASKS.md
In previous offerings, it was not clear to some students that they needed to have 5 unique CSS rules in each of their CSS files. In this update, I made the word "each" bold and added a parenthetical phrase where the 5 rules are mentioned a second time. Hopefully, this is clearer. Closes #20.