Remove the .faculty-container div
Previously, the .faculty-container div contained
the #faculty, #validation, and #style divs. This caused a bit of
confusion: people assumed that .faculty-container was only supposed to
contain #faculty.

This commit removes .faculty-container entirely. While it would
certainly be possible to keep .faculty-container around under a
different name, not using it at all is a little simpler.

This commit does not yet update any CSS files to reflect this change.

This addresses issues #41 and #43 in the
UMM-CSci-3601/3601-lab1_HTML-CSS repo.