Remove .faculty-container from style1.css
This shouldn't affect style1's layout: the positioning
of #faculty, #validation, and #style is now controlled by the body tag,
which has exactly the same flex-box rules that .faculty-container used
to have. (This means, for example, that #faculty and #validation will
still be displayed in a column, horizontally and vertically centered.)

This addresses issues #41 and #43 in the
UMM-CSci-3601/3601-lab1_HTML-CSS repo.